Clean up, prepare submission

+++ triad/src/main/java/io/github/ardoco/triad/pipeline/Enrichment.java
@@ -1,29 +1,32 @@
+/* Licensed under MIT 2025. */
 package io.github.ardoco.triad.pipeline;
+
+import java.io.IOException;
+import java.util.Map;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import io.github.ardoco.triad.ir.ArtifactsCollection;
 import io.github.ardoco.triad.ir.IRModel;
 import io.github.ardoco.triad.ir.SimilarityMatrix;
 import io.github.ardoco.triad.model.Project;
 import io.github.ardoco.triad.util.EnrichmentUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.util.Map;
 
 public class Enrichment {
     private static final Logger logger = LoggerFactory.getLogger(Enrichment.class);
 
     private final Project project;
     private final IRModel irModel;
-    private final SimilarityMatrix sourceToIntermediateSim;   // rows = sources, cols = intermediates
-    private final SimilarityMatrix targetToIntermediateSim;   // rows = targets, cols = intermediates
+    private final SimilarityMatrix sourceToIntermediateSim; // rows = sources, cols = intermediates
+    private final SimilarityMatrix targetToIntermediateSim; // rows = targets, cols = intermediates
 
-    public Enrichment(Project project,
-                      IRModel irModel,
-                      SimilarityMatrix sourceToIntermediateSim,
-                      SimilarityMatrix intermediateToTargetSim, // Not used but retained for signature parity
-                      SimilarityMatrix targetToIntermediateSim) {
+    public Enrichment(
+            Project project,
+            IRModel irModel,
+            SimilarityMatrix sourceToIntermediateSim,
+            SimilarityMatrix intermediateToTargetSim,
+            SimilarityMatrix targetToIntermediateSim) {
         this.project = project;
         this.irModel = irModel;
         this.sourceToIntermediateSim = sourceToIntermediateSim;
@@ -34,16 +37,18 @@ public class Enrichment {
         Map<String, Map<String, Integer>> intermediateBitermMap =
                 EnrichmentUtils.getBitermFrequencyMap(project.getIntermediateArtifacts());
 
-        Map<String, Map<String, Integer>> srcEnrichBiterms =
-                EnrichmentUtils.selectNeighborConsensualBiterms(project.getSourceArtifacts(), intermediateBitermMap, sourceToIntermediateSim);
+        Map<String, Map<String, Double>> srcEnrichBiterms = EnrichmentUtils.selectNeighborConsensualBiterms(
+                project.getSourceArtifacts(), intermediateBitermMap, sourceToIntermediateSim);
         EnrichmentUtils.debugEnrichmentStats("SRC+", project.getSourceArtifacts(), srcEnrichBiterms);
 
-        Map<String, Map<String, Integer>> tgtEnrichBiterms =
-                EnrichmentUtils.selectNeighborConsensualBiterms(project.getTargetArtifacts(), intermediateBitermMap, targetToIntermediateSim);
+        Map<String, Map<String, Double>> tgtEnrichBiterms = EnrichmentUtils.selectNeighborConsensualBiterms(
+                project.getTargetArtifacts(), intermediateBitermMap, targetToIntermediateSim);
         EnrichmentUtils.debugEnrichmentStats("TGT+", project.getTargetArtifacts(), tgtEnrichBiterms);
 
-        ArtifactsCollection extendedSources = EnrichmentUtils.createExtendedCollection(project.getSourceArtifacts(), srcEnrichBiterms, "SRC+");
-        ArtifactsCollection extendedTargets = EnrichmentUtils.createExtendedCollection(project.getTargetArtifacts(), tgtEnrichBiterms, "TGT+");
+        ArtifactsCollection extendedSources =
+                EnrichmentUtils.createExtendedCollection(project.getSourceArtifacts(), srcEnrichBiterms, "SRC+");
+        ArtifactsCollection extendedTargets =
+                EnrichmentUtils.createExtendedCollection(project.getTargetArtifacts(), tgtEnrichBiterms, "TGT+");
 
         SimilarityMatrix s1 = irModel.Compute(extendedSources, new ArtifactsCollection(project.getTargetArtifacts()));
         SimilarityMatrix s2 = irModel.Compute(new ArtifactsCollection(project.getSourceArtifacts()), extendedTargets);
